requirements.txt
# Core dependencies
sentence-transformers>=2.2.0
PyPDF2>=2.0.0
docx2txt>=0.8
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
numpy>=1.20.0
torch>=1.9.0
transformers>=4.15.0
pillow>=9.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.24.0

# Development dependencies
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.0
mypy>=0.930
isort>=5.10.1
